working enemy movement (NEEDS great refactor)

+++ src/BodyRectangle.cpp
@@ -1,6 +1,6 @@
 #include "BodyRectangle.hpp"
 
-BodyRectangle::BodyRectangle(int x, int y, int width, int height) {
+BodyRectangle::BodyRectangle(float x, float y, int width, int height) {
 	this->x = x;
 	this->y = y;
 	this->w = width;
